combine s3 e2e into addons e2e
