+++ e2e/addons/hello/server.js
@@ -6,8 +6,8 @@ const AWS = require('aws-sdk');
 // Constants
 const PORT = 80;
 
-// Create db connection pool
-const secret = process.env.MYSTORAGE_SECRET
+// Create db connection pool to the Aurora cluster.
+const secret = process.env.MYCLUSTER_SECRET
 const secretJSON = JSON.parse(secret)
 const Pool = require('pg').Pool
 const pool = new Pool({
@@ -18,10 +18,14 @@ const pool = new Pool({
     port: secretJSON['port'],
 })
 
+// Get bucket name from environment variable.
+const s3 = new AWS.S3({apiVersion: '2006-03-01'});
+const bucketName = process.env.MYBUCKET_NAME
+
 // App
 const app = express();
 app.get('/', (req, res) => {
-    res.send('Hello World');
+    res.status(200).send('Hello World');
 });
 
 app.get('/databases/:dbName', function(req, res, next) {
@@ -37,6 +41,21 @@ app.get('/databases/:dbName', function(req, res, next) {
             }
         }
         res.status(404)
+    })
+})
+
+app.get('/peeks3', function(req, res, next) {
+    let bucketParams = {
+        Bucket: bucketName,
+    }
+
+    s3.listObjects(bucketParams, function(err, data) {
+        if (err) {
+            res.status(400).send(err);
+            return;
+        }
+
+        res.status(200).json({'name': data.Name});
     })
 })
