--- e2e/s3/hello/.dockerignore
@@ -1,1 +1,0 @@
-node_modules
