--- e2e/s3/hello/Dockerfile
@@ -1,17 +1,0 @@
-FROM node:14
-
-# Create app directory
-WORKDIR /usr/src/app
-
-# Install app dependencies
-# A wildcard is used to ensure both package.json AND package-lock.json are copied
-# where available (npm@5+)
-COPY package*.json ./
-
-RUN npm install
-
-COPY . .
-
-EXPOSE 80
-
-CMD [ "node", "server.js" ]
