--- e2e/s3/hello/package.json
@@ -1,15 +1,0 @@
-{
-  "name": "docker_web_app",
-  "version": "1.0.0",
-  "description": "Node.js on Docker",
-  "author": "First Last <first.last@example.com>",
-  "main": "server.js",
-  "scripts": {
-    "start": "node server.js"
-  },
-  "dependencies": {
-    "aws-sdk": ">= 2.249.1",
-    "express": "^4.16.1",
-    "pg": "^8.5.1"
-  }
-}
